feat: ImageDataManager 생성 및 request 구성 메서드 분리
