+++ OpenMarket/OpenMarket/Model/Network/NetworkManager.swift
@@ -93,4 +93,37 @@ struct NetworkManager {
         }
         task.resume()
     }
+    
+    func configureRequest(_ boundary: String) -> URLRequest? {
+        guard let url = NetworkRequest.postProduct.url else {
+            return nil
+        }
+        var request = URLRequest(url: url)
+        
+        request.httpMethod = "POST"
+        request.setValue("3595be32-6941-11ed-a917-b17164efe870",
+                         forHTTPHeaderField: "identifier")
+        request.setValue("multipart/form-data; boundary=\(boundary)",
+                         forHTTPHeaderField: "Content-Type")
+        
+        return request
+    }
+    
+    func configureRequestBody(_ product: PostProduct, _ imageData: Data, _ boundary: String) -> Data? {
+        let encoder = JSONEncoder()
+        encoder.keyEncodingStrategy = .convertToSnakeCase
+        guard let productData = try? encoder.encode(product) else {
+            return nil
+        }
+        
+        var data = Data()
+        data.appendString("--\(boundary)\r\n")
+        data.appendString("Content-Disposition: form-data; name=\"params\"\r\n\r\n")
+        data.append(productData)
+        data.appendString("\r\n")
+        data.append(imageData)
+        data.appendString("\r\n--\(boundary)--\r\n")
+        
+        return data
+    }
 }
